Set up CMake build
This allows building with CMake instead of Eclipse. The reasoning behind
this is to make the code more easily portable to other architectures,
and to move away from being dependent on Eclipse.

I would like to get to the point that the AT90USB646 and AT90USB1286
bootloader versions are compiled simultaneously, but I'm not quite ready
to figure out how to do that yet.